move short name to ReactionTypes object
